Rename .h files to .hpp and update docs.
