+++ docs/arrays.hpp
