+++ test/pairs_and_sequences.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <vector>
 #include <utility>
-#include "unit_test_framework.h"
+#include "unit_test_framework.hpp"
 
 using namespace std;
