+++ test/test_name_conflict_file_with_tests.cpp
@@ -1,4 +1,4 @@
-#include "unit_test_framework.h"
+#include "unit_test_framework.hpp"
 
 // We have a function with the same name in
 // test_name_conflict_file_with_non_static_func.cpp.
