+++ test/test_special_asserts.cpp
@@ -1,4 +1,4 @@
-#include "unit_test_framework.h"
+#include "unit_test_framework.hpp"
 
 #include <iostream>
 #include <cassert>
