+++ unit_test_framework.hpp
@@ -1,5 +1,5 @@
-#ifndef UNIT_TEST_FRAMEWORK_H
-#define UNIT_TEST_FRAMEWORK_H
+#ifndef UNIT_TEST_FRAMEWORK_HPP
+#define UNIT_TEST_FRAMEWORK_HPP
 
 #include <map>
 #include <utility>
@@ -648,4 +648,4 @@ void assert_almost_equal(double first, double second, double precision,
     throw TestFailure(reason.str(), line_number, assertion_text);
 }
 
-#endif  // UNIT_TEST_FRAMEWORK_H
+#endif  // UNIT_TEST_FRAMEWORK_HPP
